requirements.txt
fastapi[all]
gunicorn
pydantic
packaging
python-dotenv
PyYAML
requests
uvicorn
web3
